requirements.txt
streamlit
ultralytics
opencv-python-headless
streamlit-webrtc
av
Pillow
numpy